Remove DiscoveryEngine
